+++ zscaler/zpa/pra_credential_pool.py
@@ -296,8 +296,8 @@ class PRACredentialPoolAPI(APIClient):
         if error:
             return (None, response, error)
 
-        if response is None:
-            return (PRACredentialPoolController({"id": pool_id}), None, None)
+        if response is None or not response.get_body():
+            return (PRACredentialPoolController({"id": pool_id}), response, None)
 
         try:
             result = PRACredentialPoolController(self.form_response_body(response.get_body()))