+++ zscaler/zpa/saml_attributes.py
@@ -273,8 +273,8 @@ class SAMLAttributesAPI(APIClient):
         if error:
             return (None, response, error)
 
-        if response is None:
-            return (SAMLAttribute({"id": attribute_id}), None, None)
+        if response is None or not response.get_body():
+            return (SAMLAttribute({"id": attribute_id}), response, None)
 
         try:
             result = SAMLAttribute(self.form_response_body(response.get_body()))