+++ zscaler/zpa/server_groups.py
@@ -295,8 +295,8 @@ class ServerGroupsAPI(APIClient):
         if error:
             return (None, response, error)
 
-        if response is None:
-            return (ServerGroup({"id": group_id}), None, None)
+        if response is None or not response.get_body():
+            return (ServerGroup({"id": group_id}), response, None)
 
         try:
             result = ServerGroup(self.form_response_body(response.get_body()))